backend/requirements.txt
# app
fastapi==0.55.1
uvicorn==0.11.3
pydantic==1.4
email-validator==1.1.1
python-multipart==0.0.5
# db
databases[postgresql]==0.4.2
SQLAlchemy==1.3.16
alembic==1.4.2
psycopg2-binary

# auth
pyjwt==2.0.1
passlib[bcrypt]==1.7.2

# dev
pytest==6.2.1
pytest-asyncio==0.14.0
httpx==0.16.1
asgi-lifespan==1.0.1

# crypto
pillow==8.2.0
PyCryptodome==3.12.0
